requirements.txt
# PDF工具箱依赖
# GUI框架
PySide6>=6.6.0

# PDF处理
pypdf>=6.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pikepdf>=8.0.0
pdfplumber>=0.10.0
pdf2docx>=0.5.0
PyMuPDF>=1.23.0

# 图片处理
Pillow>=10.0.0

# 文档处理
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
reportlab>=4.0.0

# 编码检测
chardet>=5.0.0

# 日志
loguru>=0.7.0

# Windows COM支持 (可选，仅Windows需要)
# pywin32>=306; sys_platform == 'win32'